# Add bitpacked output support to the Aarch64 optimised int16 kernel.