--- a/larq_compute_engine/core/bgemm_kernels_common.h
+++ b/larq_compute_engine/core/bgemm_kernels_common.h
@@ -2,14 +2,15 @@
 #define COMPUTE_EGNINE_TFLITE_KERNELS_BGEMM_KERNELS_COMMON_H_
 
 #include "larq_compute_engine/core/bconv2d_output_transform.h"
+#include "larq_compute_engine/core/types.h"
 #include "ruy/kernel_common.h"
 
 using namespace ruy;
 
-using compute_engine::core::OutputTransform;
-
-using compute_engine::core::bitpacking_bitwidth;
-using compute_engine::core::TBitpacked;
+namespace ce = compute_engine;
+using ce::core::bitpacking_bitwidth;
+using ce::core::OutputTransform;
+using ce::core::TBitpacked;
 
 // Our version of `ruy::MulParams`; The original is in `ruy/mul_params.h`.
 // We simply use our `OutputTransform` struct.
@@ -52,11 +53,6 @@ struct BinaryKernelParams {
   const TBitpacked* lhs_base_ptr;
   const TBitpacked* rhs_base_ptr;
   DstScalar* dst_base_ptr;
-  // post_mutiply and post_activation_bias are currently float
-  // in order to accomodate for batchnorm scales
-  // Later this might be changed to the int8 system of multipliers+shifts
-  const float* post_activation_multiplier;
-  const float* post_activation_bias;
   std::int32_t start_row;
   std::int32_t start_col;
   std::int32_t last_row;
@@ -67,42 +63,57 @@ struct BinaryKernelParams {
   std::int32_t rhs_stride;
   std::int32_t dst_stride;
   std::int32_t depth;
-  std::int32_t clamp_min;
-  std::int32_t clamp_max;
-  DstScalar dst_tmp_buf[LhsCols * RhsCols];
+  OutputTransform<DstScalar> output_transform;
+  DstScalar dst_tmp_buf[LhsCols * RhsCols];  // Used for float or int8 output
 };
 
 template <typename AccumScalar, typename DstScalar, int LhsCols, int RhsCols>
 inline void MakeBinaryKernelParams(
-    const PMat<TBitpacked>& lhs, const PMat<TBitpacked>& rhs,
-    const BinaryMulParams<AccumScalar, DstScalar>& spec, int start_row,
+    const PMat<TBitpacked>& lhs, const PMat<TBitpacked>& rhs, int start_row,
     int start_col, int end_row, int end_col, Mat<DstScalar>* dst,
+    const BinaryMulParams<AccumScalar, DstScalar>& spec,
     BinaryKernelParams<DstScalar, LhsCols, RhsCols>* params) {
-  const int depth = lhs.layout.rows;
   RUY_DCHECK_EQ(start_row % LhsCols, 0);
   RUY_DCHECK_EQ(start_col % RhsCols, 0);
   RUY_DCHECK_EQ(end_row % LhsCols, 0);
   RUY_DCHECK_EQ(end_col % RhsCols, 0);
+  if (std::is_same<DstScalar, TBitpacked>::value) {
+    RUY_DCHECK_EQ(start_row % 8, 0);
+    RUY_DCHECK_EQ(end_row % 8, 0);
+  }
 
   params->lhs_base_ptr = lhs.data + start_row * lhs.layout.stride;
   params->rhs_base_ptr = rhs.data + start_col * rhs.layout.stride;
-  params->dst_base_ptr =
-      dst->data.get() + start_col * dst->layout.stride + start_row;
-
-  params->post_activation_multiplier = spec.output_transform.multiplier;
-  params->post_activation_bias = spec.output_transform.bias;
+  if (std::is_same<DstScalar, TBitpacked>::value) {
+    // When writing bitpacked output with multiple threads, the start/end row
+    // will not necessarily be aligned to a TBitpacked boundary (though they are
+    // guaranteed to be aligned to a byte boundary). For example, start_row
+    // could be 8, in which case we'd be writing into the second byte of each
+    // TBitpacked value. Hence the need for char pointer arithmetic.
+    params->dst_base_ptr =
+        (DstScalar*)((char*)(dst->data.get() +
+                             start_col * ce::core::GetBitpackedSize(
+                                             dst->layout.stride)) +
+                     start_row / 8);
+  } else {
+    params->dst_base_ptr =
+        dst->data.get() + start_col * dst->layout.stride + start_row;
+  }
   params->start_row = start_row;
   params->start_col = start_col;
   params->last_row = end_row - LhsCols;
   params->last_col = end_col - RhsCols;
-  params->lhs_stride = sizeof(TBitpacked) * lhs.layout.stride;
-  params->rhs_stride = sizeof(TBitpacked) * rhs.layout.stride;
-  params->dst_stride = sizeof(DstScalar) * dst->layout.stride;
-  params->depth = depth;
-  params->clamp_min = spec.output_transform.clamp_min;
-  params->clamp_max = spec.output_transform.clamp_max;
   params->dst_rows = dst->layout.rows;
   params->dst_cols = dst->layout.cols;
+  params->lhs_stride = sizeof(TBitpacked) * lhs.layout.stride;
+  params->rhs_stride = sizeof(TBitpacked) * rhs.layout.stride;
+  params->dst_stride =
+      sizeof(DstScalar) * (std::is_same<DstScalar, TBitpacked>::value
+                               ? ce::core::GetBitpackedSize(dst->layout.stride)
+                               : dst->layout.stride);
+  params->depth = lhs.layout.rows;
+  params->output_transform =
+      spec.output_transform;  // This is a four word copy, but that's okay.
 
   RUY_DCHECK_LT(params->last_row, params->dst_rows);
   RUY_DCHECK_LT(params->last_col, params->dst_cols);