--- a/larq_compute_engine/tflite/kernels/bconv2d.cc
+++ b/larq_compute_engine/tflite/kernels/bconv2d.cc
@@ -415,17 +415,15 @@ inline TfLiteStatus EvalChooseKernelType(TfLiteContext* context,
                                          TfLiteBConv2DParams* params) {
   if (kernel_type == KernelType::kRuyOptimized) {
 #if RUY_PLATFORM_ARM_64
-    // On 64 bit Arm only there is an optimised kernel for 16-bit accumulators
-    // and float output. It is safe to use this without risk of overflow as long
-    // as the maximum value of the convolution (filter height * filter width *
-    // input channels, plus some overhead to account for potential padding) is
-    // less than 2^16. We will almost always take this path: for a 3x3 filter
-    // there would need to be > 7000 input channels to present an overflow risk.
+    // On 64 bit Arm only there is an optimised kernel with 16-bit accumulators.
+    // It is safe to use this without risk of overflow as long as the maximum
+    // value of the convolution (filter height * filter width * input channels,
+    // plus some overhead to account for potential padding) is less than 2^16.
+    // We will almost always take this path: for a 3x3 filter there would need
+    // to be > 7000 input channels to present an overflow risk.
     const int depth =
         params->filter_height * params->filter_width * params->channels_in;
-    if ((std::is_same<DstScalar, float>::value ||
-         std::is_same<DstScalar, std::int8_t>::value) &&
-        depth + 512 < 1 << 16) {
+    if (depth + 512 < 1 << 16) {
       EvalOpt<std::int16_t, DstScalar>(context, node, params);
       return kTfLiteOk;
     }