## What do these changes do?

This PR adds support for writing bitpacked output to the optimised assembly Aarch64 kernel with 16-bit accumulators.

Bitpacked output remains unsupported by the Aarch64 fallback optimised kernel with 32-bit accumulators, and the Arm32 optimised kernel. The reason for this is that the Aarch64 int16 kernel has a 8x4 kernel layout that writes eight output channels in parallel, whereas the others have a 4x4 kernel layout that writes only four channels in parallel. With eight channels we can always pack the bits into a byte, whereas four channels would require something more complicated (especially to support multi-threading).

## How Has This Been Tested?

CI. The 'big' kernel tests for Aarch64 and Arm32 pass locally.

## Benchmark Results

To benchmark this change, I've taken the binary alexnet model from Larq Zoo, modified it to use one-padding instead of zero-padding, and run the LCE benchmark tool (with `num_runs=250`) on my Android phone, a OnePlus 6, reporting the average latency and standard deviation:

| Model                             | Baseline       | PR            |
|-----------------------------------|----------------|---------------|
| Binary AlexNet (with one-padding) | 111.42 +- 0.60 | 64.95 +- 0.14 |

Note that the binary alexnet model includes large binary dense layers at the end of the model which are not currently accelerated by LCE. Thus, the performance improvement for the bconv op alone will be greater than the overall latency suggests. 

## Related issue number

N/A.